pyproject.toml
[project]
name = "synthetic-data-generator"
version = "0.1.0"
description = "Adaptable synthetic data generation system"
authors = [{name = "María Benavente"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "cohere>=4.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
    "tiktoken>=0.5.0",
    "transformers>=4.13.0",
    "torch>=2.0.0",
    "protobuf>=3.20.0",
    "sentencepiece>=0.1.99",
    "pandas>=2.3.3",
    "jsonschema>=4.25.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.ruff.format]
quote-style = "double"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
